udpate test config

+++ fixtures/download.sh
@@ -1,3 +1,5 @@
 #!/usr/bin/env bash
 
-curl https://homepages.cae.wisc.edu/~ece533/images/girl.png -o fixtures/girl.png
+if [ ! -f "fixtures/girl.png" ]; then
+  curl https://homepages.cae.wisc.edu/~ece533/images/girl.png -o fixtures/girl.png
+fi
